Move methods to abstract class
