+++ mod-erm-usage-counter41/src/main/java/org/olf/erm/usage/counter41/csv/mapper/AbstractCounterReport.java
@@ -3,6 +3,7 @@ package org.olf.erm.usage.counter41.csv.mapper;
 import com.google.common.collect.Iterables;
 import java.io.IOException;
 import java.io.StringWriter;
+import java.math.BigInteger;
 import java.time.LocalDate;
 import java.time.YearMonth;
 import java.time.format.DateTimeFormatter;
@@ -13,11 +14,16 @@ import java.util.List;
 import java.util.Locale;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
+import org.niso.schemas.counter.MetricType;
+import org.niso.schemas.counter.PerformanceCounter;
 import org.niso.schemas.counter.Report;
+import org.niso.schemas.counter.ReportItem;
 import org.olf.erm.usage.counter41.Counter4Utils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.supercsv.io.CsvListWriter;
+import org.supercsv.io.dozer.CsvDozerBeanWriter;
+import org.supercsv.io.dozer.ICsvDozerBeanWriter;
 import org.supercsv.prefs.CsvPreference;
 
 public abstract class AbstractCounterReport {
@@ -36,19 +42,19 @@ public abstract class AbstractCounterReport {
 
   public abstract String getDescription();
 
-  public String[] createHeader() {
+  private String[] createHeader() {
     Stream<String> header = Arrays.stream(getHeader());
     Stream<String> months = yearMonths.stream().map(ym -> ym.format(formatter));
     return Stream.concat(header, months).toArray(String[]::new);
   }
 
-  public String[] createFieldMapping() {
+  private String[] createFieldMapping() {
     Stream<String> mapping = Arrays.stream(getFieldMapping());
     Stream<String> months = Collections.nCopies(yearMonths.size(), "itemPerformance").stream();
     return Stream.concat(mapping, months).toArray(String[]::new);
   }
 
-  public String createReportHeader() {
+  private String createReportHeader() {
     StringWriter stringWriter = new StringWriter();
 
     try (CsvListWriter csvListWriter =
@@ -73,12 +79,94 @@ public abstract class AbstractCounterReport {
     return stringWriter.toString();
   }
 
+  public String[] createTotals() {
+    return null;
+  }
+
+  public abstract void writeItems(ICsvDozerBeanWriter writer) throws IOException;
+
+  public final String getSinglePublisher() {
+    List<String> uniquePublishers =
+        report.getCustomer().get(0).getReportItems().stream()
+            .map(ReportItem::getItemPublisher)
+            .distinct()
+            .collect(Collectors.toList());
+    LOG.info("Found {} publishers: {}", uniquePublishers.size(), uniquePublishers);
+    return (uniquePublishers.size() == 1) ? uniquePublishers.get(0) : null;
+  }
+
+  public final String getSinglePlatform() {
+    List<String> uniquePlatforms =
+        report.getCustomer().get(0).getReportItems().stream()
+            .map(ReportItem::getItemPlatform)
+            .distinct()
+            .collect(Collectors.toList());
+    LOG.info("Found {} platforms: {}", uniquePlatforms.size(), uniquePlatforms);
+    return (uniquePlatforms.size() == 1) ? uniquePlatforms.get(0) : null;
+  }
+
+  public final String bigIntToStringOrNull(BigInteger bigint) {
+    return (bigint == null) ? null : String.valueOf(bigint);
+  }
+
+  public final BigInteger getPeriodMetricTotal(MetricType metricType, YearMonth month) {
+    return report.getCustomer().get(0).getReportItems().stream()
+        .flatMap(ri -> ri.getItemPerformance().stream())
+        .filter(
+            ip ->
+                month == null
+                    || (ip.getPeriod()
+                            .getBegin()
+                            .toGregorianCalendar()
+                            .toZonedDateTime()
+                            .toLocalDate()
+                            .equals(month.atDay(1))
+                        && ip.getPeriod()
+                            .getEnd()
+                            .toGregorianCalendar()
+                            .toZonedDateTime()
+                            .toLocalDate()
+                            .equals(month.atEndOfMonth())))
+        .flatMap(m -> m.getInstance().stream())
+        .filter(pc -> pc.getMetricType().equals(metricType))
+        .map(PerformanceCounter::getCount)
+        .reduce(BigInteger::add)
+        .orElse(null);
+  }
+
   /**
    * Returns a CSV representation of the report
    *
    * @return CSV formatted String
    */
-  public abstract String toCSV();
+  public String toCSV() {
+    StringWriter stringWriter = new StringWriter();
+
+    stringWriter.append(createReportHeader());
+
+    // and the formatted stuff with CsvDozerBeanWriter
+    try (ICsvDozerBeanWriter beanWriter =
+        new CsvDozerBeanWriter(stringWriter, CsvPreference.STANDARD_PREFERENCE)) {
+
+      // configure the mapping from the fields to the CSV columns
+      beanWriter.configureBeanMapping(ReportItem.class, createFieldMapping());
+
+      // write the header
+      beanWriter.writeHeader(createHeader());
+
+      // write totals line
+      String[] totals = createTotals();
+      if (totals != null) beanWriter.writeHeader(totals);
+
+      writeItems(beanWriter);
+
+      beanWriter.flush();
+      return stringWriter.toString();
+    } catch (IOException e) {
+      LOG.error(e.getMessage(), e);
+      return null;
+    }
+  };
 
   /**
    * Returns a seamless list of every month that falls within the reporting period.
